ElasticSearch aufräumen/verbessern

+++ app/grails-app/services/de/laser/DeletionService.groovy
@@ -924,8 +924,10 @@ class DeletionService {
         RestHighLevelClient esclient = ESWrapperService.getClient()
 
         try {
-            DeleteRequest request = new DeleteRequest(es_index, id)
-            DeleteResponse deleteResponse = esclient.delete(request, RequestOptions.DEFAULT);
+            if(ESWrapperService.testConnection()) {
+                DeleteRequest request = new DeleteRequest(es_index, id)
+                DeleteResponse deleteResponse = esclient.delete(request, RequestOptions.DEFAULT);
+            }
             esclient.close()
         }
         catch(Exception e) {
